backend/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.13
django-colorfield==0.4.5
django-cors-headers==3.10.1
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==3.9.2
flake8-isort==4.1.1
freezegun==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==20.9
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.9.2
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==3.8.0
pytest-freezegun==0.4.2
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
testfixtures==6.18.3
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.4.1
